+++ ocaml-lsp-server/src/custom_requests/custom_request.mli
@@ -10,3 +10,4 @@ module Type_enclosing = Req_type_enclosing
 module Wrapping_ast_node = Req_wrapping_ast_node
 module Get_documentation = Req_get_documentation
 module Type_search = Req_type_search
+module Merlin_jump = Req_merlin_jump